// src/@types/api/types/chats.ts
export interface ApiRestrictionReason {
  reason: string;
  text: string;
}